move tests to abstract class
