+++ tests/Feature/PendingScopedFeatureInteraction/DatabaseDriverPendingScopedFeatureInteraction.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Feature\PendingScopedFeatureInteraction;
+
+use Illuminate\Support\Facades\Config;
+
+class DatabaseDriverPendingScopedFeatureInteraction extends AbstractPendingScopedFeatureInteraction
+{
+    protected function setDriver(): void
+    {
+        Config::set('pennant.default', 'database');
+    }
+}
